pyproject.toml
[project]
name = "draw-a-holiday-calendar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.56.0",
    "pillow>=10.0.0",
    "openai>=1.0.0",
]
